backend/pyproject.toml
[tool.poetry]
name = "video-risk-analyzer"
version = "0.1.0"
description = "SNS Video Risk Analyzer Backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = {extras = ["redis"], version = "^5.3.6"}
python-multipart = "^0.0.6"
boto3 = "^1.34.0"
google-cloud-speech = "^2.24.0"
google-cloud-videointelligence = "^2.13.0"
google-cloud-aiplatform = "^1.40.0"
ffmpeg-python = "^0.2.0"
sse-starlette = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
fakeredis = "^2.21.0"
moto = {extras = ["s3"], version = "^5.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__pycache__/*"]
